#884 - Partial fix
