+++ src/routes/two-factor-authentication/remove-secret.ts
@@ -1,26 +1,18 @@
-import Joi from 'joi';
 import { ControllerGeneratorOptionsWithoutClientOrSupplier } from '../../comms/controller';
 
-interface Request {
-  query: {
-    email: string;
-  };
-}
-
+type Request = Record<string, undefined> | undefined;
+type EffectiveRequest = Record<string, undefined>;
 type Response = void;
 
 const controllerGeneratorOptions: ControllerGeneratorOptionsWithoutClientOrSupplier = {
   method: 'post',
   path: '/remove-secret',
-  query: Joi.object().keys({
-    email: Joi.string().required().example('test@test.com'),
-  }).required(),
   right: {},
 };
 
 export {
   controllerGeneratorOptions,
   Request,
-  Request as EffectiveRequest,
+  EffectiveRequest,
   Response,
 };
