+++ src/routes/two-factor-authentication/routes.ts
@@ -2,12 +2,10 @@ import * as validatePassword from './validate-password';
 import * as validateCode from './validate-code';
 import * as removeSecret from './remove-secret';
 import * as getQRCode from './get-qr-code';
-import * as getSettings from './get-settings';
 
 export {
   validatePassword,
   validateCode,
   removeSecret,
   getQRCode,
-  getSettings,
 };
