+++ src/routes/user/index.ts
@@ -4,6 +4,7 @@ import * as find from './find';
 import * as get from './get';
 import * as update from './update';
 import Comms from '../../comms';
+
 import controllerGenerator, { Result } from '../../comms/controller';
 import TableController from '../../comms/table-controller';
